T.ApexLM.AzureAI/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
azure-ai-textanalytics==5.3.0
python-dotenv==1.0.0
azure-core==1.29.5